+++ src/provenanceJSONcommon.h
@@ -1,17 +1,21 @@
+/* SPDX-License-Identifier: GPL-2.0 */
 /*
-*
-* Author: Thomas Pasquier <thomas.pasquier@bristol.ac.uk>
-*
-* Copyright (C) 2015-2016 University of Cambridge
-* Copyright (C) 2016-2017 Harvard University
-* Copyright (C) 2017-2018 University of Cambridge
-* Copyright (C) 2018-202O University of Bristol
-*
-* This program is free software; you can redistribute it and/or modify
-* it under the terms of the GNU General Public License version 2, as
-* published by the Free Software Foundation.
-*
-*/
+ * Copyright (C) 2015-2016 University of Cambridge,
+ * Copyright (C) 2016-2017 Harvard University,
+ * Copyright (C) 2017-2018 University of Cambridge,
+ * Copyright (C) 2018-2021 University of Bristol,
+ * Copyright (C) 2021-2022 University of British Columbia
+ *
+ * Author: Thomas Pasquier <tfjmp@cs.ubc.ca>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 2, as
+ * published by the Free Software Foundation; either version 2 of the License,
+ * or (at your option) any later version.
+ */
+
+#include <unistd.h>
+
 #define MAX_JSON_BUFFER_EXP     13
 #define MAX_JSON_BUFFER_LENGTH  ((1 << MAX_JSON_BUFFER_EXP)*sizeof(uint8_t))
 #define BUFFER_LENGTH (MAX_JSON_BUFFER_LENGTH-strnlen(buffer, MAX_JSON_BUFFER_LENGTH))
@@ -142,4 +146,19 @@ static inline void __add_machine_id(uint32_t value, bool comma){
   strncat(buffer, "\"cf:", BUFFER_LENGTH);
   strncat(buffer, utoa(value, tmp, DECIMAL), BUFFER_LENGTH);
   strncat(buffer, "\"", BUFFER_LENGTH);
+}
+
+static inline void __add_name_id(union prov_identifier* name_id, bool comma){
+  char *name;
+  int i;
+  if (name_id->node_id.type == 0)
+    return;
+  for (i = 0; i < 5; i++) {
+    name = name_id_to_str(name_id);
+    if (name != NULL)
+      break;
+    usleep(50000);
+  }
+  if (name != NULL)
+    __add_string_attribute("cf:name", name, comma);
 }